requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
Pillow>=10.2.0
opencv-python>=4.9.0.80
matplotlib>=3.8.0
scipy>=1.12.0
tqdm>=4.66.0